requirements.txt
langchain
langchain-classic
langchain-core
langchain-groq
langchain-ollama
langchain-huggingface
langchain-community
faiss-cpu  # or faiss-gpu
python-dotenv
torch
sentence-transformers
streamlit

